remove compiler warnings with options -Wextra -pedantic global floating point variables are now initialized with -atomic option since INIT_MAIN has not been called.

+++ algorithms/ForwardNumerics/BaseIntegerCommon.h
@@ -79,7 +79,7 @@ class UnsignedBaseStoreTraits {
       {  memcpy(target, source, count*sizeof(unsigned int)); }
    static int sizeBaseInBits() { return 8*sizeof(unsigned); }
    static int minCellsCountToStoreBits(int bitsNumber)
-      {  return (bitsNumber + 8*sizeof(unsigned) - 1) / (8*sizeof(unsigned)); } 
+      {  return (int) ((bitsNumber + 8*sizeof(unsigned) - 1) / (8*sizeof(unsigned))); } 
 
    static void swapArray(unsigned int* target, unsigned int* source, int count)
       {  unsigned int temp;
